Crate chain_rpc

Modules

client

Extended JsonRpcClient abstraction This module contains custom implementation of ethers::providers::JsonRpcClient which allows usage of non-reqwest based HTTP clients.

errors

indexer

Indexer specific trait implementation (HoprIndexerRpcOperations)

rpc

General purpose high-level RPC operations implementation (HoprRpcOperations)

Structs

BlockWithLogs

Structure containing filtered logs that all belong to the same block.

Log

A type containing selected fields from the eth_getLogs RPC calls. This is further restricted to already mined blocks.

LogFilter

Represents a filter to extract logs containing specific contract events from a block.

PendingTransaction

Represents a pending transaction that can be eventually resolved until confirmation, which is done by polling the respective RPC provider. The polling interval and number of confirmations are defined by the underlying provider.

TransactionReceipt

Contains some selected fields of a receipt for a transaction that has been already included into the blockchain.

ZeroRetryPolicy

Performs no retries.

Enums

RetryAction

Indicates what retry action should be taken, as result of a RetryPolicy implementation.

TypedTransaction

The TypedTransaction enum represents all Ethereum transaction types.

Traits

HoprIndexerRpcOperations

Trait with RPC provider functionality required by the Indexer.

HoprRpcOperations

Trait defining general set of operations an RPC provider must provide to the HOPR node.

HttpPostRequestor

Abstraction for HTTP client that perform HTTP POST with serializable request data.

RetryPolicy

Simple retry policy trait

Functions

create_eip1559_transaction

Short-hand for creating new EIP1559 transaction object.