Trait chain_rpc::RetryPolicy

pub trait RetryPolicy<E> {
    // Provided method
    fn is_retryable_error(
        &self,
        _err: &E,
        _retry_number: u32,
        _retry_queue_size: u32
    ) -> RetryAction { ... }
}

Simple retry policy trait

Provided Methods

fn is_retryable_error( &self, _err: &E, _retry_number: u32, _retry_queue_size: u32 ) -> RetryAction

Indicates whether a client should retry the request given the last error, current number of retries of this request and the number of other requests being retried by the client at this time.